+++ Proyecto/Web/app/classes/logics/gastos_imprevistos.py
@@ -12,7 +12,7 @@ class Gestor_gastos_imprevistos:
     def logic(self) -> None:
         try:
             df_gastos_imprevistos = pd.read_csv("/usr/src/app/app/classes/logics/data/Controldegastos/data_gastos_imprevistos.csv")
-            st.subheader("Gastos imprevistos")
+            st.subheader("Gastos Imprevistos")
             st.dataframe(df_gastos_imprevistos)
             st.markdown("<br>", unsafe_allow_html=True)
             st.divider()
@@ -20,79 +20,59 @@ class Gestor_gastos_imprevistos:
             if df_gastos_imprevistos.empty:
                 self._add_gastos(df_gastos_imprevistos)
             else:
-                tab_mostrar_gastos,tab_agregar_gastos ,tab_modificacion_gastos = st.tabs(["Mi información","Agregar nuevo gasto" ,"Modificación de información"])
+                tab_mostrar_gastos,tab_agregar_gastos ,tab_modificacion_gastos = st.tabs(["Mis Gastos","Agregar Gasto" ,"Modificación de Gastos"])
                 with tab_mostrar_gastos:
                     self._displey_gastos(df_gastos_imprevistos)
                 with tab_agregar_gastos:
                     self._add_gastos(df_gastos_imprevistos)
                 with tab_modificacion_gastos:
                     self._modify_gastos(df_gastos_imprevistos)
-
+                    
         except FileNotFoundError:
             st.error("Archivo CSV no encontrado. Por favor, verifíca la ruta.")
             df_gastos_imprevistos = pd.DataFrame(
                 columns=[
                     "Categoria",
                     "Servicio",
-                    "Gasto mensual",
-                    "Descripción",
-                    "Comentario",
+                    "Descripcion",
+                    "Costo",
                 ]
             )
             st.dataframe(df_gastos_imprevistos)
 
-    def _add_gastos(self, df_gastos_imprevistos):
+    def _add_gastos(self, df_gastos_imprevistos:dict) -> None:
         tipo_categoria = st.selectbox(
-        "Categoria",
-        ("Vivienda", "Salud","Alimentos","Entretenimiento","Transporte"),
+            "Categoría",
+            ("Vivienda", "Salud","Alimentos","Entretenimiento","Transporte"),
         )
         servicios = {
-            "Vivienda": ["Alquiler", "Luz", "Agua", "Gas", "Internet"],
-            "Salud": ["Seguro medico", "Medicamentos", "Agua", "Gimnasio"],
-            "Alimentos": ["Comida", "Bebidas", "Productos frescos", "Comida rapida"],
-            "Entretenimiento": ["Netflix", "Spotify", "Amazon Prime", "Disney +", "Entretenimiento general"],
-            "Transporte": ["Gasolina", "Transporte público", "Mantenimiento del vehículo"],
+            "Vivienda": ["Reparaciones", "Mudanza"],
+            "Salud": ["Medicamentos", "Emergencia médica", "Emergencia quirúrgica"],
+            "Alimentos": ["Comida", "Bebidas", "Comida rápida"],
+            "Entretenimiento": ["Juegos", "Entradas para eventos", "Música"],
+            "Transporte": ["Mantenimiento del vehículo", "Tramite vehicular", "Movilidad por aplicación"],
         }
         tipo_servicio = st.selectbox("Servicio", servicios[tipo_categoria])
-        gasto_mensual = st.text_input("Gasto mensual")
-        detalles_del_servicio = {
-            "Alquiler": "Costo de vivienda",
-            "Luz": "Energía eléctrica",
-            "Agua": "Suministro de agua potable",
-            "Gas": "Suministro de gas doméstico",
-            "Internet": "Conexión a internet",
-            "Seguro medico": "Plan de seguro de salud",
-            "Medicamentos": "Gastos en medicinas",
-            "Gimnasio": "Suscripción mensual al gimnasio",
-            "Comida": "Compras de alimentos para el hogar",
-            "Bebidas": "Consumo de bebidas en general",
-            "Productos frescos": "Frutas, verduras y carnes",
-            "Comida rapida": "Gastos en cadenas de comida",
-            "Netflix": "Suscripción mensual de streaming",
-            "Spotify": "Suscripción mensual de música",
-            "Amazon Prime": "Suscripción mensual para streaming y envíos",
-            "Entretenimiento general": "Gastos en ocio y actividades recreativas",
-            "Gasolina": "Costo del combustible",
-            "Transporte público": "Gastos en autobuses o trenes",
-            "Mantenimiento del vehículo": "Reparaciones y servicios"
-        }
-        descripcion = detalles_del_servicio[tipo_servicio]
-        comentario = st.text_input("Comentario (opcional)")
+        descripcion = st.text_input("Descripcion (Opcional)")
+        costo = st.text_input("Costo")
+        st.write("")
         if st.button("Agregar Gasto"):
-            new_record = {
-                "Categoria": tipo_categoria,
-                "Servicio": tipo_servicio ,
-                "Gasto mensual": gasto_mensual,
-                "Descripción": descripcion,
-                "Comentario": comentario,
-            }
-            df_gastos_imprevistos = df_gastos_imprevistos.append(new_record, ignore_index=True)
-            df_gastos_imprevistos.to_csv("/usr/src/app/app/classes/logics/data/Controldegastos/data_gastos_imprevistos.csv", mode="w", index=False)
-            st.success("Gasto agregado exitosamente.")
-            time.sleep(3)
-            st.experimental_rerun()
+            if not str(costo).replace('.', '', 1).isdigit() or float(costo) < 0:
+                st.error("El costo debe ser un número positivo.")
+            else:
+                new_record = {
+                    "Categoria": tipo_categoria,
+                    "Servicio": tipo_servicio ,
+                    "Descripcion": descripcion,
+                    "Costo": float(costo),
+                }
+                df_gastos_imprevistos = df_gastos_imprevistos.append(new_record, ignore_index=True)
+                df_gastos_imprevistos.to_csv("/usr/src/app/app/classes/logics/data/Controldegastos/data_gastos_imprevistos.csv", mode="w", index=False)
+                st.success("Gasto agregado exitosamente.")
+                time.sleep(3)
+                st.experimental_rerun()
 
-    def _displey_gastos(self, df_gastos_imprevistos):
+    def _displey_gastos(self, df_gastos_imprevistos:dict) -> None:
         st.markdown(
             """
             <link href="https://cdn.jsdelivr.net/npm/bootstrap-icons/font/bootstrap-icons.css" rel="stylesheet">
@@ -100,11 +80,11 @@ class Gestor_gastos_imprevistos:
             unsafe_allow_html=True
         )
         for index, fila in df_gastos_imprevistos.iterrows():
+            n_gasto = index + 1
             tipo_categoria = fila['Categoria']
             tipo_servicio = fila['Servicio']
-            gasto_mensual = fila['Gasto mensual']
-            descripcion = fila['Descripción']
-            comentario = fila['Comentario']
+            descripcion = fila['Descripcion']
+            costo = fila['Costo']
             iconos_servicio = {
             "Vivienda": "house-heart",
             "Salud": "heart-pulse",
@@ -112,84 +92,91 @@ class Gestor_gastos_imprevistos:
             "Entretenimiento": "emoji-laughing",
             "Transporte": "truck-front-fill",
             }
-
             icono_gasto = iconos_servicio.get(tipo_categoria)
-
-            st.write(f"Gasto para: {tipo_categoria}")
+            st.subheader(f"Gasto de {tipo_categoria}")
             st.markdown(
                 f"""
                 <div style="
-                border: 2px solid white; 
-                border-radius: 10px; 
-                padding: 10px; 
-                background-color: #151516;
-                width: 100%; 
-                max-width: 800px; 
-                margin: auto;">
-                <p style='color: white;'> 
-                <i class="bi bi-{icono_gasto}" style="margin-right: 8px;"></i></p>
-                <p style='color: white;'> Servicio: {tipo_servicio}</p>
-                <p style='color: white;'> Gasto mensual: {gasto_mensual}</p>
-                <p style='color: white;'> Descripcion: {descripcion}</p>
-                <p style='color: white;'> Comentario: {comentario}</p>
+                    border: 2px solid white; 
+                    border-radius: 10px; 
+                    padding: 10px; 
+                    background-color: #151516;
+                    width: 100%; 
+                    max-width: 800px; 
+                    margin: auto;">
+                    <p style='color: white;'> 
+                    <i class="bi bi-{icono_gasto}" style="margin-right: 8px;"></i></p>
+                    <p style='color: white;'> Número de Gasto: {n_gasto}</p>
+                    <p style='color: white;'> Servicio: {tipo_servicio}</p>
+                    <p style='color: white;'> Descripción: {descripcion}</p>
+                    <p style='color: white;'> Costo: {costo}</p>
                 </div>
                 """,
                 unsafe_allow_html=True
             )
-            st.write("") 
-            st.write("") 
+            st.write("")
 
-    def _modify_gastos(self, df_gastos_imprevistos):
-        st.subheader("Modificar Gastos Existentes")
+    def _modify_gastos(self, df_gastos_imprevistos:dict) -> None:
+        st.subheader("Modificar Gastos Imprevistos Existentes")
         if not df_gastos_imprevistos.empty:
-            selected_servicio = st.selectbox(
-                "Selecciona un contacto para modificar",
-                df_gastos_imprevistos['Servicio'],
+            opciones = list(range(1, len(df_gastos_imprevistos) + 1))
+            selected_usuario = st.selectbox(
+                "Selecciona el número del gasto a modificar", opciones
             )
-            if selected_servicio is not None:
-                st.write(f" {selected_servicio}") 
-                selected_record = df_gastos_imprevistos[df_gastos_imprevistos['Servicio'] == selected_servicio].iloc[0]
-                tipo_categoria = selected_record['Categoria']
-                tipo_servicio = selected_record['Servicio']
-                gasto_mensual = st.text_input(
-                    "Gasto mensual",
-                    value=selected_record["Gasto mensual"],
-                    key=",Gasto mensual",
+            selected_index = selected_usuario - 1
+            if selected_index is not None:
+                selected_record = df_gastos_imprevistos.loc[selected_index]
+                tipo_categoria = st.selectbox(
+                    "Categoría",
+                    ["Vivienda", "Salud", "Alimentos", "Entretenimiento", "Transporte"],
+                    index=["Vivienda", "Salud", "Alimentos", "Entretenimiento", "Transporte"].index(selected_record["Categoria"]),
+                    key="Categoria",
                 )
+                st.write("")
+                tipo_servicio = st.text_input(
+                    "Servicio",
+                    value=selected_record["Servicio"],
+                    key=",Servicio",
+                )
+                st.write("")
                 descripcion = st.text_input(
                     "Descripción",
-                    value=selected_record["Descripción"],
+                    value=selected_record["Descripcion"],
                     key=",Descripcion",
                 )
-                comentario = st.text_input(
-                    "Comentario",
-                    value=selected_record["Comentario"],
-                    key=",Comentario",
+                costo = st.text_input(
+                    "Costo",
+                    value=selected_record["Costo"],
+                    key=",Costo",
                 )
-                st.write("") 
-                st.write("") 
-                if st.button("Guardar Cambios", key="save_button"):
-                    df_gastos_imprevistos.loc[df_gastos_imprevistos['Servicio'] == selected_servicio, 'Categoria'] = tipo_categoria
-                    df_gastos_imprevistos.loc[df_gastos_imprevistos['Servicio'] == selected_servicio, 'Servicio'] = tipo_servicio
-                    df_gastos_imprevistos.loc[df_gastos_imprevistos['Servicio'] == selected_servicio, 'Gasto mensual'] = gasto_mensual
-                    df_gastos_imprevistos.loc[df_gastos_imprevistos['Servicio'] == selected_servicio, 'Descripción'] = descripcion
-                    df_gastos_imprevistos.loc[df_gastos_imprevistos['Servicio'] == selected_servicio, 'Comentario'] = comentario
-                    df_gastos_imprevistos.to_csv(
-                        "/usr/src/app/app/classes/logics/data/Controldegastos/data_gastos_imprevistos.csv",
-                        mode="w",
-                        index=False,
-                    )
-                    st.success("Registro modificado exitosamente.")
-                    time.sleep(2)
-                    st.experimental_rerun()
-                    st.write("") 
-                    st.write("")
-                    if st.button("Eliminar gasto", key="save_button2"):
+                st.write("")
+                col1, col2 = st.columns(2)
+                with col1:
+                    if st.button("Guardar Cambios", key="save_button"):
+                        df_gastos_imprevistos.at[selected_index, "Categoria"] = tipo_categoria
+                        df_gastos_imprevistos.at[selected_index, "Servicio"] = tipo_servicio
+                        df_gastos_imprevistos.at[selected_index, "Descripcion"] = descripcion
+                        if not str(costo).replace('.', '', 1).isdigit() or float(costo) < 0:
+                            st.error("El costo debe ser un número positivo.")
+                        else:
+                            df_gastos_imprevistos.at[selected_index, "Costo"] = float(costo)
+                            df_gastos_imprevistos.to_csv(
+                                "/usr/src/app/app/classes/logics/data/Controldegastos/data_gastos_imprevistos.csv",
+                                mode="w",
+                                index=False,
+                            )
+                        st.success("Gasto modificado exitosamente.")
+                        time.sleep(3) 
+                        st.experimental_rerun() 
+                st.write("")
+                with col2:
+                    if st.button("Eliminar Gasto", key="save_button2"):
                         csv_file = '/usr/src/app/app/classes/logics/data/Controldegastos/data_gastos_imprevistos.csv'
-                        df_gastos_imprevistos = df_gastos_imprevistos[df_gastos_imprevistos['Servicio'] != selected_servicio] 
-                        df_gastos_imprevistos.to_csv(csv_file, index=False)  
+                        df_gastos_imprevistos = df_gastos_imprevistos.drop(index=selected_index)
+                        df_gastos_imprevistos = df_gastos_imprevistos.reset_index(drop=True)
+                        df_gastos_imprevistos.to_csv(csv_file, index=False)
                         st.success("Gasto eliminado exitosamente.")
-                        time.sleep(3)
+                        time.sleep(3) 
                         st.experimental_rerun()
         else:
-            st.error("No hay registros disponibles para modificar.")
+            st.error("No existen gastos disponibles para modificar.")
